Protect against protocol desyncs that can occur after a malformed put command
The put command must have a data payload to trigger the desync.  On a
malformed put with data, the server sends *two* response lines: a
BAD_FORMAT for the command and then an UNKNOWN_COMMAND for the received
data.  If the sysread() in _interact() happens to return just the first
line, _interact() will stop reading and the UNKNOWN_COMMAND line will
remain for the next _interact() command to mistakenly read as its
response.  How long the desync persists depends on the subsequent
commands and their read/write patterns as well as how much sysread()
happens to return.

Another solution is to refactor the line-wise read in _interact() into a
_recv_one_line() function and then call the new function on a put error
(ignoring the contents, or warning if the contents aren't the expected
UNKNOWN_COMMAND).  This would resync the protocol without disconnecting
but feels more fragile than simply resetting the connection.  Other
error handling in the client uses a forced reset, so I did the same.

A test is included which will induce failure (on most runs, though not
necessarily every run) without the client patch.

See also <https://github.com/kr/beanstalkd/issues/141>.